requirements.txt
python-telegram-bot==20.3
openai==0.28.0
python-dotenv==1.0.0
pytest==7.4.0 